current status: strange ball

+++ Makefile
@@ -1,8 +1,9 @@
 src = $(wildcard ./src/*.cc)
 header = $(wildcard ./include/)
 target = main.out
-temp_image = out.ppm
-target_image = out.jpg
+temp_image = ./target/out.ppm
+target_image = ./target/out.jpg
+
 CC = g++
 CCFLAGS = -std=c++17 -O2 -Wall -Werror
 
@@ -11,18 +12,21 @@ $(target): $(src)
 
 .PHONY: clean
 clean:
-	rm -rf ${temp_image};
-	rm -rf $(target);
-	rm -rf ${target_image};
+ifndef keep-temp
+	rm -rf ${temp_image}
+endif
+	rm -rf $(target) ${target_image}
 
 .PHONY: run
 run:
-	make && ./$(target) && python ./src/main.py && code ./${target_image};
-	# rm -rf ${temp_image};
-	rm -rf $(target);
+	make && ./$(target) && python -c "import cv2;cv2.imwrite('${target_image}', cv2.imread('${temp_image}'))" && code ./${target_image}
+ifndef keep-temp
+	rm -rf ${temp_image}
+endif
+	rm -rf $(target)
 
 .PHONY: all
 all:
-	make clean;
-	make run;
+	make clean
+	make run
 	date +"%F %T"
